challenge8/requirements.txt
semantic-kernel>=1.0.0
openai>=1.0.0
pyodbc>=5.0.0
python-dotenv>=1.0.0
